+++ pkg/oto.go
@@ -1,11 +1,11 @@
 package oto
 
 import (
-	"fmt"
-	"time"
 	"bytes"
 	"context"
+	"fmt"
 	"os/exec"
+	"time"
 
 	"github.com/Bl4omArchie/fme"
 	"github.com/Bl4omArchie/simple"
@@ -20,13 +20,13 @@ type Instance struct {
 	Database       *gorm.DB
 	ParamsSchema   map[string]fme.Schema
 	TemporalClient client.Client
-	Workers map[string]WorkerItem
+	Workers        map[string]WorkerItem
 }
 
 type WorkerItem struct {
-	WorkerID string
-	Worker worker.Worker
-	OutputError chan(error)
+	WorkerID    string
+	Worker      worker.Worker
+	OutputError chan (error)
 }
 
 type Config struct {
@@ -38,6 +38,7 @@ type Config struct {
 	TemporalHost      string `env:"TEMPORAL_HOST,required"`
 	TemporalNamespace string `env:"TEMPORAL_NAMESPACE,required"`
 }
+
 
 func NewInstanceOto(envPath string) (*Instance, error) {
 	cfg, err := simple.GetEnv[Config](envPath)
@@ -71,8 +72,14 @@ func NewInstanceOto(envPath string) (*Instance, error) {
 
 // === Add data ===
 
-func (i *Instance) AddExecutable(name, version, executablePath, description string) error {
-	exec := models.NewExecutable(name, version, executablePath, description)
+// Add and save a new executable
+//
+// name			: executable name
+// version		: executable version
+// path			: absolute path of the executable on your machine
+// description	: add a few words to describe your executable
+func (i *Instance) AddExecutable(name, version, path, description string) error {
+	exec := models.NewExecutable(name, version, path, description)
 	if err := i.Database.Save(exec).Error; err != nil {
 		return fmt.Errorf("failed to save Executable: %w", err)
 	}
@@ -80,25 +87,36 @@ func (i *Instance) AddExecutable(name, version, executablePath, description stri
 	return nil
 }
 
-func (i *Instance) AddParameter(ctx context.Context, execTag, flag, description string, requiresRoot, requiresValue bool, valueType models.ValueType, Require, InterfersWith []string, s *fme.Schema) error {
+
+// Add and save a new parameter
+//
+// execName		: executable name, ( executable name - version ) i.e : "nmap - 7.94"
+// flag			: parameter flag
+// description	: add a few words to describe your parameter
+// requiresRoot : if this parameter require to use root rigth, set to true, else to false
+// requiresRoot : if this parameter require a value beside it, set to true, else to false
+// valueType 	: whathever the parameter require a value or not, precise here the value he can support. Type None if this parameter can't support any values.
+// requires		: add flag of parameter this one depends on.
+// interfers	: add flag of parameter this one can't be mixed to.
+func (i *Instance) AddParameter(ctx context.Context, execName, flag, description string, requiresRoot, requiresValue bool, valueType models.ValueType, requires, interfers []string, s *fme.Schema) error {
 	// Retrieve executable
-	exec, err := models.FetchExecutable(ctx, i.Database, "tag", execTag)
+	exec, err := models.FetchExecutable(ctx, i.Database, "tag", execName)
 	if err != nil {
 		return err
 	}
 
 	// Retrieve dependencies
-	RequireToSave, err := models.FetchFlagParameters(ctx, i.Database, "flag", Require)
+	requireToSave, err := models.FetchFlagParameters(ctx, i.Database, "flag", requires)
 	if err != nil {
 		return err
 	}
-	InterfersWithToSave, err := models.FetchFlagParameters(ctx, i.Database, "flag", InterfersWith)
+	InterfersWithToSave, err := models.FetchFlagParameters(ctx, i.Database, "flag", interfers)
 	if err != nil {
 		return err
 	}
 
 	// Verify dependencies correctness
-	for _, depends := range RequireToSave {
+	for _, depends := range requireToSave {
 		if ok, err := s.Require(flag, depends.Flag); ok == false {
 			return fmt.Errorf("incorrect `requirement` dependency : %v", err)
 		}
@@ -109,15 +127,21 @@ func (i *Instance) AddParameter(ctx context.Context, execTag, flag, description 
 		}
 	}
 
-	param := models.NewParameter(flag, description, exec, requiresRoot, requiresValue, valueType, InterfersWithToSave, RequireToSave)
+	param := models.NewParameter(flag, description, exec, requiresRoot, requiresValue, valueType, InterfersWithToSave, requireToSave)
 	if err := i.Database.Save(param).Error; err != nil {
 		return fmt.Errorf("failed to parameter : %w", err)
 	}
 	return nil
 }
 
-func (i *Instance) AddCommand(ctx context.Context, execID, cmdName, description string, flags []string, s *fme.Schema) error {
-	exec, err := models.FetchExecutable(ctx, i.Database, "tag", execID)
+
+// Add and save a new command
+//
+// execName		: executable tag, ( executable name - version ) i.e : "nmap - 7.94"
+// flag			: parameter flag
+// description	: add a few words to describe your parameter
+func (i *Instance) AddCommand(ctx context.Context, execTag, cmdName, description string, root bool, flags []string, s *fme.Schema) error {
+	exec, err := models.FetchExecutable(ctx, i.Database, "tag", execTag)
 	if err != nil {
 		return err
 	}
@@ -127,19 +151,32 @@ func (i *Instance) AddCommand(ctx context.Context, execID, cmdName, description 
 		return err
 	}
 
+	// TODO : verify root flags for parameters
+
 	// FME : check if the given flags are valid before ingestion
 	_, err = s.ValidateCombination(flags)
 	if err != nil {
 		return err
 	}
 
-	cmd := models.NewCommand(cmdName, description, exec, flagsToSave)
+	cmd := models.NewCommand(cmdName, description, exec, root, flagsToSave)
 	if err := i.Database.Save(cmd).Error; err != nil {
 		return fmt.Errorf("failed to save command: %w", err)
 	}
 	return nil
 }
 
+
+// Add and save a new job
+//
+// execName		: executable tag, ( executable name - version ) i.e : "nmap - 7.94"
+// flag			: parameter flag
+// description	: add a few words to describe your parameter
+// requiresRoot : if this parameter require to use root rigth, set to true, else to false
+// requiresRoot : if this parameter require a value beside it, set to true, else to false
+// valueType 	: whathever the parameter require a value or not, precise here the value he can support. Type None if this parameter can't support any values.
+// requires		: add flag of parameter this one depends on.
+// interfers	: add flag of parameter this one can't be mixed to.
 func (i *Instance) AddJob(ctx context.Context, cmdName, jobName string, flagValues map[string]string) error {
 	cmd, err := models.FetchCommand(ctx, i.Database, "name", cmdName)
 	if err != nil {
@@ -162,6 +199,7 @@ func (i *Instance) AddJob(ctx context.Context, cmdName, jobName string, flagValu
 	return nil
 }
 
+
 func (i *Instance) ImportParameters(ctx context.Context, filename string, s *fme.Schema) error {
 	params, err := simple.LoadFile[models.ParameterRaw](filename, -1, true)
 	if err != nil {
@@ -169,7 +207,7 @@ func (i *Instance) ImportParameters(ctx context.Context, filename string, s *fme
 	}
 
 	for _, p := range params {
-		err := i.AddParameter(ctx, p.ExecutableTag, p.Flag, p.Description, p.RequiresRoot, p.RequiresValue, p.ValueType, p.RequireIDs, p.InterferIDs, s)
+		err := i.AddParameter(ctx, p.ExecutableName, p.Flag, p.Description, p.RequiresRoot, p.RequiresValue, p.ValueType, p.RequireIDs, p.InterferIDs, s)
 		if err != nil {
 			return fmt.Errorf("failed to add parameter %s: %w", p.Flag, err)
 		}
@@ -177,8 +215,28 @@ func (i *Instance) ImportParameters(ctx context.Context, filename string, s *fme
 	return nil
 }
 
+func (i *Instance) AddRoutine(ctx context.Context, routineName string, jobsMap map[models.Action]string) error {
+	var routine *models.Routine
+	var routineMap map[models.Action]models.Job
+
+	for action, jobName := range jobsMap {
+		job, err := models.FetchJob(ctx, i.Database, "name", jobName)
+		if err != nil {
+			return fmt.Errorf("couldn't find job : %s. %v", job, err)
+		}
+		routineMap[action] = *job
+	}
+
+	routine = models.NewRoutine(routineName, routineMap, make([]models.Output, 0))
+
+	if err := i.Database.Save(routine).Error; err != nil {
+		return fmt.Errorf("failed to save job command: %w", err)
+	}
+	return nil
+}
+
 // Tmp function for demo only. This will be erased when Temporal we'll be fully integrated.
-func (i *Instance) RunJobDemo(ctx context.Context, jobName string) (*JobOutput, error) {
+func (i *Instance) RunJobDemo(ctx context.Context, jobName string) (*models.Output, error) {
 	job, err := models.FetchJob(ctx, i.Database, "name", jobName)
 	if err != nil {
 		return nil, err
@@ -209,18 +267,17 @@ func (i *Instance) RunJobDemo(ctx context.Context, jobName string) (*JobOutput, 
 	cmd.Stderr = &stderr
 
 	err = cmd.Run()
-	return &JobOutput{
+	return &models.Output{
 		Stdout: stdout.String(),
 		Stderr: stderr.String(),
 	}, err
 }
 
-
 // == FME ===
-func (i *Instance) AddExecutableSchema(ctx context.Context, execTag string) (*fme.Schema, error) {
+func (i *Instance) AddExecutableSchema(ctx context.Context, execName string) (*fme.Schema, error) {
 	s := fme.NewSchema()
 
-	exec, err := models.FetchExecutable(ctx, i.Database, "tag", execTag)
+	exec, err := models.FetchExecutable(ctx, i.Database, "tag", execName)
 	if err != nil {
 		return nil, err
 	}
@@ -244,16 +301,15 @@ func (i *Instance) AddExecutableSchema(ctx context.Context, execTag string) (*fm
 	if err != nil {
 		return nil, err
 	}
-	i.ParamsSchema[execTag] = *s
+	i.ParamsSchema[execName] = *s
 	return s, nil
 }
-
 
 // === Temporal ===
 func (i *Instance) NewWorkerItem(workerID string) WorkerItem {
 	return WorkerItem{
-		WorkerID: workerID,
-		Worker: worker.New(i.TemporalClient, workerID, worker.Options{}),
+		WorkerID:    workerID,
+		Worker:      worker.New(i.TemporalClient, workerID, worker.Options{}),
 		OutputError: make(chan error),
 	}
 }
@@ -269,7 +325,7 @@ func (i *Instance) AddWorker(workerID string) error {
 
 	acts := &Activities{DB: i.Database}
 
-	w.Worker.RegisterWorkflow(WorkflowRunJob)
+	w.Worker.RegisterWorkflow(WorkflowRunRoutine)
 	w.Worker.RegisterActivity(acts.RunJob)
 
 	go func() {
@@ -292,18 +348,30 @@ func (i *Instance) StopWorker(workerID string) error {
 	return nil
 }
 
-func (i *Instance) RunJobWorkflow(ctx context.Context, jobName string) (*JobOutput, error) {
+func (i *Instance) ConfigWorker(workerID string) error {
+	// TODO
+
+	return nil
+}
+
+func (i *Instance) ConfigWorkflow(workerID string) error {
+	// TODO
+
+	return nil
+}
+
+func (i *Instance) RunWorkflow(ctx context.Context, workerID, jobName string) (*models.Output, error) {
 	workflowOptions := client.StartWorkflowOptions{
 		ID:        fmt.Sprintf("job-%s-%d", jobName, time.Now().UnixNano()),
-		TaskQueue: "oto-tasks",
+		TaskQueue: workerID,
 	}
 
-	handle, err := i.TemporalClient.ExecuteWorkflow(ctx, workflowOptions, WorkflowRunJob, jobName)
+	handle, err := i.TemporalClient.ExecuteWorkflow(ctx, workflowOptions, WorkflowRunRoutine, jobName)
 	if err != nil {
 		return nil, err
 	}
 
-	var result JobOutput
+	var result models.Output
 	if err := handle.Get(ctx, &result); err != nil {
 		return nil, err
 	}